student-list/simple_api/requirements.txt
flask==2.0.0
flask_httpauth==4.1.0
python-dotenv==0.14.0
Werkzeug==2.0.0
fluent-logger